refactor "ChainTo" to instance method "Then"
